feature(datastore): implement ds as a plugin
This makes it available from anywhere you have access to an editorstate
by simply importing the citationKey and requesting its state.

This is especially useful for pluginfunctions. Existing controller api
remains as before.
