+++ addon/core/prosemirror.ts
@@ -1,12 +1,9 @@
 import { Command, EditorState, Plugin, Transaction } from 'prosemirror-state';
 import { EditorView, NodeViewConstructor } from 'prosemirror-view';
 import {
-  Attrs,
   DOMParser as ProseParser,
   DOMSerializer,
-  Mark,
   MarkType,
-  Node as PNode,
   Schema,
 } from 'prosemirror-model';
 import { baseKeymap, selectAll, toggleMark } from 'prosemirror-commands';
@@ -20,13 +17,16 @@ import { tracked } from '@glimmer/tracking';
 import { dropCursor } from 'prosemirror-dropcursor';
 import MapUtils from '../utils/map-utils';
 import { createLogger, Logger } from '../utils/logging-utils';
-import { filter, objectValues } from 'iter-tools';
 import {
   ProseStore,
-  proseStoreFromParse,
   ResolvedPNode,
 } from '@lblod/ember-rdfa-editor/utils/datastore/prose-store';
 import { ReferenceManager } from '@lblod/ember-rdfa-editor/utils/reference-manager';
+import {
+  datastore,
+  datastoreKey,
+} from '@lblod/ember-rdfa-editor/plugins/datastore';
+import { unwrap } from '@lblod/ember-rdfa-editor/utils/option';
 
 export type WidgetLocation =
   | 'toolbarMiddle'
@@ -59,21 +59,13 @@ interface ProsemirrorArgs {
 export default class Prosemirror {
   view: EditorView;
   @tracked _state: EditorState;
-  @tracked datastore: ProseStore;
   @tracked widgets: Map<WidgetLocation, InternalWidgetSpec[]> = new Map();
   root: Element;
   baseIRI: string;
   pathFromRoot: Node[];
   schema: Schema;
-  private readonly tag: (node: ResolvedPNode) => string;
-  private readonly attributes: (node: ResolvedPNode) => Attrs;
-  private readonly isText: (node: ResolvedPNode) => boolean;
 
   private logger: Logger;
-  private children: (
-    schema: Schema,
-    refman: ProseReferenceManager
-  ) => (resolvedNode: ResolvedPNode) => Iterable<ResolvedPNode>;
 
   constructor({
     target,
@@ -87,11 +79,13 @@ export default class Prosemirror {
   }: ProsemirrorArgs) {
     this.logger = createLogger(this.constructor.name);
     this.root = target;
+    this.pathFromRoot = getPathFromRoot(this.root, false);
     this.baseIRI = baseIRI;
     this.schema = schema;
     this._state = EditorState.create({
       doc: ProseParser.fromSchema(this.schema).parse(target),
       plugins: [
+        datastore({ pathFromRoot: this.pathFromRoot, baseIRI }),
         ...plugins,
 
         dropCursor(),
@@ -107,24 +101,6 @@ export default class Prosemirror {
       attributes: { class: 'say-editor__inner say-content' },
       nodeViews: nodeViews(new ProseController(this)),
       dispatchTransaction: this.dispatch,
-    });
-    this.pathFromRoot = getPathFromRoot(this.root, false);
-    this.tag = tag(this.schema);
-    this.children = children;
-    this.attributes = attributes(this.schema);
-    this.isText = isText(this.schema);
-    const refman = new ProseReferenceManager();
-    this.datastore = proseStoreFromParse({
-      root: refman.get({ node: this._state.doc }),
-      textContent,
-      tag: this.tag,
-      children: this.children(this.schema, refman),
-      attributes: this.attributes,
-      isText: this.isText,
-      getParent: getParent(refman),
-
-      pathFromDomRoot: this.pathFromRoot,
-      baseIRI,
     });
     this.initializeEditorWidgets(widgets);
   }
@@ -154,23 +130,6 @@ export default class Prosemirror {
 
   dispatch = (tr: Transaction) => {
     const newState = this.state.apply(tr);
-
-    if (tr.docChanged) {
-      const refman = new ProseReferenceManager();
-      this.datastore = proseStoreFromParse({
-        textContent,
-        tag: this.tag,
-        children: this.children(this.schema, refman),
-        attributes: this.attributes,
-        isText: this.isText,
-        getParent: getParent(refman),
-        root: refman.get({ node: newState.doc }),
-        pathFromDomRoot: this.pathFromRoot,
-        baseIRI: this.baseIRI,
-      });
-      this.logger(`Parsed ${this.datastore.size} triples`);
-    }
-
     this.view.updateState(newState);
     this._state = newState;
   };
@@ -246,7 +205,7 @@ export class ProseController {
   }
 
   get datastore(): ProseStore {
-    return this.pm.datastore;
+    return unwrap(datastoreKey.getState(this.pm.state));
   }
 
   get widgets() {
@@ -279,7 +238,7 @@ export class ProseController {
   }
 }
 
-class ProseReferenceManager extends ReferenceManager<
+export class ProseReferenceManager extends ReferenceManager<
   ResolvedPNode,
   ResolvedPNode
 > {
@@ -291,117 +250,4 @@ class ProseReferenceManager extends ReferenceManager<
       }
     );
   }
-}
-
-function textContent(resolvedNode: ResolvedPNode): string {
-  return resolvedNode.node.textContent;
-}
-
-function isText(schema: Schema) {
-  return function (resolvedNode: ResolvedPNode): boolean {
-    const { node } = resolvedNode;
-    if (getLinkMark(schema, node)) {
-      return false;
-    }
-    return node.isText;
-  };
-}
-
-function getLinkMark(schema: Schema, node: PNode): Mark | undefined {
-  if (!schema.marks.link) {
-    return undefined;
-  }
-  const linkMarks = filter(
-    (markType: MarkType) => markType.spec.group === 'linkmarks',
-    objectValues(schema.marks)
-  );
-  const isText = node.isText;
-  if (isText) {
-    for (const type of linkMarks) {
-      const mark = type.isInSet(node.marks);
-      if (mark) {
-        return mark;
-      }
-    }
-  }
-  return undefined;
-}
-
-function children(schema: Schema, refman: ProseReferenceManager) {
-  return function (resolvedNode: ResolvedPNode): Iterable<ResolvedPNode> {
-    let result: Iterable<ResolvedPNode>;
-    const { node, pos: resolvedPos } = resolvedNode;
-    if (node.isText) {
-      const linkMark = getLinkMark(schema, node);
-      if (linkMark) {
-        result = [
-          refman.get({
-            node: node.mark(linkMark.removeFromSet(node.marks)),
-            pos: resolvedPos,
-          }),
-        ];
-      } else {
-        result = [];
-      }
-    } else {
-      const root = resolvedPos ? resolvedPos.doc : node;
-      const rslt: ResolvedPNode[] = [];
-      node.descendants((child, relativePos) => {
-        const absolutePos = resolvedPos
-          ? resolvedPos.pos + 1 + relativePos
-          : relativePos;
-        rslt.push(
-          refman.get({
-            node: child,
-            pos: root.resolve(absolutePos),
-          })
-        );
-        return false;
-      });
-      result = rslt;
-    }
-    return result;
-  };
-}
-
-function tag(schema: Schema) {
-  return function (resolvedNode: ResolvedPNode): string {
-    const { node } = resolvedNode;
-    if (getLinkMark(schema, node)) {
-      return 'a';
-    }
-    return node.type.name;
-  };
-}
-
-function attributes(schema: Schema) {
-  return function (resolvedNode: ResolvedPNode): Record<string, string> {
-    const { node } = resolvedNode;
-    const linkMark = getLinkMark(schema, node);
-    if (linkMark) {
-      return linkMark.attrs;
-    }
-    return node.attrs;
-  };
-}
-
-function getParent(refman: ProseReferenceManager) {
-  return function (
-    resolvedNode: ResolvedPNode,
-    resolvedRoot: ResolvedPNode
-  ): ResolvedPNode | null {
-    let result: ResolvedPNode | null;
-    const { pos } = resolvedNode;
-    if (!pos) {
-      result = null;
-    } else if (pos.depth === 0) {
-      result = refman.get({ node: resolvedRoot.node });
-    } else {
-      result = refman.get({
-        node: pos.parent,
-        pos: resolvedRoot.node.resolve(pos.before(pos.depth)),
-      });
-    }
-    return result;
-  };
 }
