+++ app/commands/insert-html-command.js
@@ -1,0 +1,1 @@
+export * from '@lblod/ember-rdfa-editor/commands/insert-html-command';
