+++ app/commands/type-commands.js
@@ -1,0 +1,1 @@
+export * from '@lblod/ember-rdfa-editor/commands/type-commands';
