+++ app/plugin/placeholder.js
@@ -1,0 +1,1 @@
+export * from '@lblod/ember-rdfa-editor/plugins/placeholder';
